Add project overview and local project management
